tehtävä 4.21

+++ osa4/blogilista/controllers/blogs.js
@@ -39,6 +39,15 @@ blogsRouter.get('/', async (request, response) => {
   })
 
   blogsRouter.delete('/:id', async (request, response) => {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+    if (!(blog.user.toString() === decodedToken.id.toString())) {
+      return response.status(401).json({ error: 'no rights to remove the blog' })
+    }
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
   })
